+++ zenml/steps/preprocesser/standard_preprocesser/methods/standard_methods.py
@@ -15,7 +15,7 @@
 
 import tensorflow_transform as tft
 
-from zenml.core.steps.preprocesser.standard_preprocesser.methods import \
+from zenml.steps.preprocesser.standard_preprocesser.methods import \
     methods_nonseq_filling, methods_transform
 from zenml.utils.preprocessing_utils import MethodDescriptions
